Add files via upload
Analysis
The project is a flashcard game where the user begins with typing their name in. After that, they select what kind of Mathematical operation they would like to do, including Addition, Subtraction, Multiplication and Division. They can only select one operation per run of the program. After that, it will ask the user for the highest and lowest number of factors they would like to be involved in the math problems. Following that, they will be asked to input the number of problems they would like to solve.
It will then directly process the information and output a question to the user. If the user is correct, it will print out “Correct”, else it will say “Incorrect”. After going through the number of problems (iterations) that the user wanted, it will go the Session Summary where it will state the number of problems done, the number of correct answers, a score, the amount of time taken to complete the game along with the User’s name, the type of operation they did, and the date and time.


Design
When the program is initialized, it initiates the String name, ops, and caseName which invokes getName() method, getOperation() method and null, respectively. Following that, the integer variables maxRange, minRange, numProblem, sysAns, userAns, and userScore are initiated where the former three invoke getMaxRange, getMinRange and, getProbNum methods to get the ranges and the number of problems, respectively, while the latter three are initialized as 0.

A long startTime is used to begin counting when the application was started. Then we go into a for-loop. The loop is conditioned as, as long as the number of problems the user wanted to is not executed, the loop will keep running.
Two integers are initiated in the for-loop called rand1 and rand2, to invoke the getRand1 and getRand2 methods, which generate the random numbers for the math problem between the high and the low range that the user requested.

After this comes the case-switch. This one takes the information from the String ops to determine what kind of operation the user selected. 
If the user selected A, it will go to case Addition, where the caseName will be assigned and a question will be printed to the user asking the answer, while at the same time, the system will also calculate the factors. 

If the user selected S, it will go to case Subtraction, where the caseName will be assigned. To keep things easy for the user, assuming they are mainly children or kindergarteners, the two random values will be added in the system and then the product less random2 value will be shown as a question, thus the answer will be the random1 value. It stays within their limit.

If the user selected M, it will go to case Multiplication, where the caseName will be assigned and a question will be printed to the user asking the answer, and the system will calculate.
 
If the user selected D, it will go to the case Division, where the caseName will be assigned. To keep things easy for the user, just like subtraction, the random numbers will be multiplied and then the product divided by random2 value will be shown as a question, thus the answer will be the rand1 value. It also stays within their limit.
Each case will then break out, preventing it from moving to the next case, and move to the next part of the code with userAns, which is outside the switch, where the user’s answer will be saved. This will then move to the if-else loop.
In the if-else loop, the if-loop will have a condition where it will match the user’s answer with the system’s answer. If it is correct, it will print out “Correct” and at the same time will count up the user score with userScore++ variable. Else, it will print out “Incorrect”, and not count up the user score.

After the iterations are all done, the program will calculate the score as a percentage. Another timer will count the end time and save it. After that, the difference between the end time and started time will be calculated to give the total time in milliseconds, which will then converted to seconds.

Finally, the Session summary will be shown to the user, where it will state the number of problems done, the number of correct answers, the score, the amount of time taken to complete the game along with the User’s name, the type of operation they did, and the date and time.

Under the main method, the other methods are written, which were invoked by each of the variables described to get the above information. Each of these methods all worked the same, where they asked the user a single question, the user typed in the answer, and the values were returned. The getRand1 and getRand2 methods used the minRange and maxRange to generate random numbers, which were returned to the main method above. For rand2, we made sure that the numbers generated are never 0, as it would crash the program during the division game, using a do-while loop.

Testing.

I tested the program using different operations and a different number of questions to make sure that the program worked and if there were any crashes, I went back to the code and debugged it. I have used multiple ranges of values, including positives and negatives to see if the system works. All the operators were used thoroughly.